+++ fundbot/plugins/fund/data_source.py
@@ -1,5 +1,4 @@
 import httpx
-import requests
 import time
 import random
